llm_service/requirements.txt
fastapi
uvicorn[standard]
llama-cpp-python[server]
jinja2
python-dotenv
